# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Docs api",
        "version": "1.0.0"
    },
    "paths": {
...
        "/cart/": {
            "get": {
                "tags": [
                    "Cart"
                ],
                "summary": "Get products in the cart",
                "description": "Retrieves product information for the specified product IDs in the cart.",
                "operationId": "getCartProducts",
                "parameters": [
                    {
                        "name": "products",
                        "in": "query",
                        "description": "Array of product IDs",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer",
                                "format": "int64"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Product"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/cart/buy": {
            "post": {
                "tags": [
                    "Cart"
                ],
                "summary": "Purchase products in the cart",
                "description": "Processes a purchase transaction for the specified products by the authenticated user.",
                "operationId": "buyProducts",
                "requestBody": {
                    "description": "Request body for buying products",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "products"
                                ],
                                "properties": {
                                    "products": {
                                        "description": "Array of product IDs to purchase",
                                        "type": "array",
                                        "items": {
                                            "type": "integer",
                                            "format": "int64"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Products have been successfully purchased",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Products have been successfully purchased"
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "balance": "Insufficient funds on the balance sheet"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/download": {
            "get": {
                "tags": [
                    "Download"
                ],
                "summary": "Download a file",
                "description": "Downloads a file from the server based on the provided file path.",
                "operationId": "downloadFile",
                "parameters": [
                    {
                        "name": "path",
                        "in": "query",
                        "description": "The path of the file to download",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "File downloaded successfully",
                        "content": {
                            "application/octet-stream": {
                                "schema": {
                                    "type": "string",
                                    "format": "binary"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "File not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "File not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/payment/{currency}": {
            "post": {
                "tags": [
                    "Payment"
                ],
                "summary": "Process a payment transaction",
                "description": "Process a payment transaction",
                "operationId": "a4b0ecbcd6911ed7bd7ef872af515001",
                "parameters": [
                    {
                        "name": "currency",
                        "in": "path",
                        "description": "Currency code (e.g., USDT, BTC)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Payment process details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "unique_key": {
                                        "description": "Unique key for the transaction",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Transaction processed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "boolean"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/payment/exchange/{currency}": {
            "post": {
                "tags": [
                    "Payment"
                ],
                "summary": "Get exchange rates",
                "description": "Get exchange rates for a specific currency",
                "operationId": "2da3e19c9919124aad2a85c60784ceb1",
                "parameters": [
                    {
                        "name": "currency",
                        "in": "path",
                        "description": "Currency code (e.g., USD, EUR)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Product": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "productable_id": {
                        "type": "integer"
                    },
                    "productable_type": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        }
    }
}